MANAGER/requirements.txt
# WebSocket server
fastapi>=0.115.0
uvicorn>=0.30.0
wsproto>=1.2.0

# Configuration
PyYAML>=6.0.2

# OpenStack / API
requests>=2.32.0

# Docker
docker>=7.0.0

# Triton Inference
tritonclient[grpc,http]>=2.48.0
numpy>=1.26.0
protobuf>=5.26.0

# MinIO / S3
boto3>=1.35.0
